simulation/src/reporter/requirements.txt
environs==9.5.0
asyncio-dgram==2.1.2
PyJWT==2.4.0
aiohttp==3.8.1